PixSpectrum-main/BackEnd/requirements.txt
fastapi
uvicorn
opencv-python
numpy
scipy
python-multipart
pymongo
python-dotenv
appwrite
passlib
bcrypt
jose
gunicorn
pyjwt
pillow
cloudinary